Improve Google Lighthouse accessibility score
Improve the accessibility score of all hyperbo.la views to 100 in Google
Lighthouse.

The accessbility score previously could not be measured because text was
low-contrast.

This PR makes the following changes:

- Use `text-light` text color from bootstrap.
- Tweak navbar active page gradients to give more contrast with
  hyperbola-colored links.
- Lighten the color of timestamps by 20% from text-muted with a custom
  CSS color and class defined in `index.scss`.

This PR also includes a missed change from #27 and changes an h6 to a div in the
blog index listing.